feat(game): Changing all values from int to float
- Changing all SDL_Rect to SDL_FRect in Player.h, Game.cpp and Game.h
- Changing types from int to float in Point.h, Player.cpp and Player.h
- Changing all "%d" to "%f" of printf() in Game.cpp
- Changing SDL_RenderDrawLine to SDL_RenderDrawLineF in Game.cpp
- Changing all SDL_RenderFillRect to SDL_RenderFillRectF in Game.cpp

+++ src/Game/Player.cpp
@@ -5,7 +5,7 @@
  * @brief Implements the Player class representing a player in a 2D game.
  */
 
-Player::Player(int startX, int startY, int playerSpeed, int playerWidth, int playerHeight)
+Player::Player(float startX, float startY, float playerSpeed, float playerWidth, float playerHeight)
         : x(startX), y(startY), speed(playerSpeed), width(playerWidth), height(playerHeight) {}
 
 std::vector<Point> Player::getVertices() const {
